Clean up and eliminate unnecessary db queries.

+++ CC2013/views.py
@@ -39,7 +39,7 @@ def modify_program(program_id=None):
     db.session.commit()
 
     #flash('New entry was successfully posted')
-    return redirect('/program/{0}'.format(program.id))
+    return redirect('/program/{program.id}'.format(program=program))
 
 
 # Program Summary, and Program Edit/Delete
@@ -56,8 +56,7 @@ def program(program_id, action=None):
         elif action == 'delete':
             # Delete the given program
             # TODO: Add a confirmation dialog here
-            courses = Course.query.filter_by(program_id=program.id).all()
-            for course in courses:
+            for course in program.courses:
                 db.session.delete(course)
             db.session.delete(program)
             db.session.commit()
@@ -67,27 +66,29 @@ def program(program_id, action=None):
             abort(404)
     else:
         # No action provided, display program summary
-        courses = Course.query.filter_by(program_id=program_id).all()
         units = []
-        for course in courses:
+        for course in program.courses:
             units.extend(course.units)
         unit_ids = [unit.id for unit in units]
-        print 'courses', courses
+        print 'courses', program.courses
         print 'units', units
         print 'unit_ids', unit_ids
 
+        # Calculate the core hours for the courses in this program
         if unit_ids:
             tier1, tier2 = db.session.query(db.func.sum(Unit.tier1).label('Tier1'),
                                             db.func.sum(Unit.tier2).label('Tier2')).filter(
                                                 Unit.id.in_(unit_ids)).first()
         else:
             tier1, tier2 = 0, 0
+
+        # All core hours
         tier1_total, tier2_total = db.session.query(db.func.sum(Unit.tier1).label('Tier1'),
                                                     db.func.sum(Unit.tier2).label('Tier2')).first()
-        return render_template('program.html', program=program, courses=courses,
-                               tier1=tier1, tier2=tier2,
-                               tier1_total=tier1_total,
-                               tier2_total=tier2_total)
+        return render_template('program.html', program=program,
+                               courses=list(program.courses),
+                               tier1=tier1, tier1_total=tier1_total,
+                               tier2=tier2, tier2_total=tier2_total)
 
 
 # Course creation
@@ -101,49 +102,54 @@ def new_course(program_id):
 @app.route('/program/<int:program_id>/edit_course/<int:course_id>/', methods=['POST'])
 @app.route('/program/<int:program_id>/add_course', methods=['POST'])
 def modify_course(program_id, course_id=None):
+    # Get form data
     title = request.form['course_title'].strip()
     abbr = request.form['course_abbr'].strip()
     description = request.form['course_description'].strip()
 
-    # Query db
-    program = Program.query.filter_by(id=program_id).first_or_404()
+    # Modify the course
     if course_id:
         # Course exists, so update properties
         course = Course.query.filter_by(id=course_id).first_or_404()
+        if course.program.id != program_id:
+            abort(404)
         course.title = title
         course.abbr = abbr
         course.description = description
     else:
         # Create new course object and store in the database
+        program = Program.query.filter_by(id=program_id).first_or_404()
         course = Course(program, title, abbr, description)
         db.session.add(course)
     db.session.commit()
 
     #flash('New entry was successfully posted')
-    return redirect(url_for('course', program_id=program.id, course_id=course.id))
+    return redirect(url_for('course', program_id=program_id, course_id=course.id))
 
 
 # Add Knowledge Units and/or Learning Outcomes to a course
 @app.route('/program/<int:program_id>/course/<int:course_id>/add_<source>', methods=['POST'])
 def add_outcomes(program_id, course_id, source):
-    program = Program.query.filter_by(id=program_id).first_or_404()
     course = Course.query.filter_by(id=course_id).first_or_404()
+    if course.program.id != program_id:
+        abort(404)
 
-    # Lookup outcomes selected on the page, in the database
+    # Query outcomes selected on the page, by KU or individually
     if source == 'unit':
-        getlist = request.form.getlist('knowledge_units')
-        unit_ids = [int(item) for item in getlist]
+        units_list = request.form.getlist('knowledge_units')
+        unit_ids = [int(unit) for unit in units_list]
         units = Unit.query.filter(Unit.id.in_(unit_ids)).all()
-        course.units.extend(units)
         outcomes = Outcome.query.filter(Outcome.unit_id.in_(unit_ids)).all()
     elif source == 'outcome':
-        getlist = request.form.getlist('learning_outcomes')
-        outcome_ids = [int(item) for item in getlist]
+        units = []
+        outcomes_list = request.form.getlist('learning_outcomes')
+        outcome_ids = [int(outcome) for outcome in outcomes_list]
         outcomes = Outcome.query.filter(Outcome.id.in_(outcome_ids)).all()
     else:
         abort(404)
 
     # Add the outcomes to the course and redisplay course information
+    course.units.extend(units)
     course.outcomes.extend(outcomes)
     db.session.commit()
     return redirect(url_for('course', program_id=program_id, course_id=course_id))
@@ -152,26 +158,38 @@ def add_outcomes(program_id, course_id, source):
 # Execute unit deletion
 @app.route('/program/<int:program_id>/course/<int:course_id>/unit/<int:unit_id>/delete')
 def delete_unit(program_id, course_id, unit_id):
-    # TODO: Verify that program_id, course_id, and unit_id are related
+    # Query db and verify that program_id, course_id, and unit_id are related
     course = Course.query.filter_by(id=course_id).first_or_404()
     if course.program.id != program_id:
         abort(404)
     unit = Unit.query.filter_by(id=unit_id).first_or_404()
-    if unit in course.units:
-        course.units.remove(unit)
-        db.session.commit()
-    else:
+    if unit not in course.units:
         abort(404)
+
+    # Delete!
+    for outcome in unit.outcomes:
+        course.outcomes.remove(outcome)
+    course.units.remove(unit)
+    db.session.commit()
+
     return redirect(url_for('course', program_id=program_id, course_id=course_id))
 
 
 # Execute outcome deletion
 @app.route('/program/<int:program_id>/course/<int:course_id>/outcome/<int:outcome_id>/delete')
 def delete_outcome(program_id, course_id, outcome_id):
+    # Query db and verify that program_id, course_id, and outcome_id are related
     course = Course.query.filter_by(id=course_id).first_or_404()
+    if course.program.id != program_id:
+        abort(404)
     outcome = Outcome.query.filter_by(id=outcome_id).first_or_404()
+    if outcome not in course.outcomes:
+        abort(404)
+
+    # Delete!
     course.outcomes.remove(outcome)
     db.session.commit()
+
     return redirect(url_for('course', program_id=program_id, course_id=course_id))
 
 
@@ -179,32 +197,33 @@ def delete_outcome(program_id, course_id, outcome_id):
 @app.route('/program/<int:program_id>/course/<int:course_id>/<action>')
 @app.route('/program/<int:program_id>/course/<int:course_id>')
 def course(program_id, course_id, action=None):
-    program = Program.query.filter_by(id=program_id).first_or_404()
     course = Course.query.filter_by(id=course_id).first_or_404()
+    if course.program.id != program_id:
+        abort(404)
 
     if action:
         if action == 'edit':
             # Edit program level info
-            return render_template('edit_course.html', program=program, course=course)
+            return render_template('edit_course.html', course=course,
+                                   program=course.program)
         elif action == 'delete':
             # Delete the given course
             # TODO: Add a confirmation dialog here
             db.session.delete(course)
             db.session.commit()
-            return redirect(url_for('program', program_id=program.id))
+            return redirect(url_for('program', program_id=program_id))
         else:
             # Nonexistant action
             abort(404)
     else:
         # No action provided, display course information
-        t1 = [outcome for outcome in course.outcomes if outcome.tier == 1]
-        t2 = [outcome for outcome in course.outcomes if outcome.tier == 2]
-        elec = [outcome for outcome in course.outcomes if outcome.tier == 3]
-        course_outcomes = [t1, t2, elec]
+        course_outcomes = [filter(lambda o: o.tier == 1, course.outcomes),
+                           filter(lambda o: o.tier == 2, course.outcomes),
+                           filter(lambda o: o.tier == 3, course.outcomes)]
         areas = Area.query.order_by(Area.id).all()
         tier1 = sum([unit.tier1 for unit in course.units])
         tier2 = sum([unit.tier2 for unit in course.units])
-        return render_template('course.html', program=program, course=course,
+        return render_template('course.html', course=course,
                                course_outcomes=course_outcomes, areas=areas,
                                tier1=tier1, tier2=tier2)
 
@@ -212,21 +231,33 @@ def course(program_id, course_id, action=None):
 # AJAX here...
 @app.route('/json')
 def get_json():
+    # Get form data...
     area_id = request.args.get('area_id', '')
     course_id = int(request.args.get('course_id', '0'))
+
+    # Get listbox data
     if area_id:
+        # A KA was selected, get all relevent KUs
         course = Course.query.filter_by(id=course_id).first()
         print course
         if course.units:
             units = Unit.query.filter_by(area_id=area_id).filter(~Unit.id.in_([u.id for u in course.units])).all()
         else:
             units = Unit.query.filter_by(area_id=area_id).all()
-        junits = [{'id': u.id, 'text': u.text, 'tier1': u.tier1, 'tier2': u.tier2} for u in units]
+        junits = [{'id': u.id,
+                   'text': u.text,
+                   'tier1': u.tier1,
+                   'tier2': u.tier2} for u in units]
         return json.dumps(junits)
     else:
+        # A KU was selected, get all relevent LOs
         unit_id = request.args.get('unit_id', 0, type=int)
         outcomes = Outcome.query.filter_by(unit_id=unit_id).order_by(Outcome.number).all()
-        joutcomes = [{'id': o.id, 'text': o.text, 'tier': ['', 'Tier 1', 'Tier 2', 'Elective'][o.tier], 'mastery': o.mastery} for o in outcomes]
+        tiers = ['', 'Tier 1', 'Tier 2', 'Elective']
+        joutcomes = [{'id': o.id,
+                      'text': o.text,
+                      'tier': tiers[o.tier],
+                      'mastery': o.mastery} for o in outcomes]
         return json.dumps(joutcomes)
